+++ core/class/jMQTTCmd.class.php
@@ -63,22 +63,16 @@ class jMQTTCmd extends cmd {
 	 */
 	private function eventNewCmd($reload=false) {
 		$eqLogic = $this->getEqLogic();
-		$eqLogic->log('info', 'Command added ' . $this->getType() . ' ' . $this->getLogName());
-
+		$eqLogic->log('info', sprintf(__('Commande %s #%s# ajoutée', __FILE__), $this->getType(), $this->getHumanName()));
 		// Advise the desktop page (jMQTT.js) that a new command has been added
-		event::add('jMQTT::cmdAdded',
-			array('eqlogic_id' => $eqLogic->getId(), 'eqlogic_name' => $eqLogic->getName(),
-				'cmd_name' => $this->getName(), 'reload' => $reload));
+		event::add('jMQTT::cmdAdded', array('eqlogic_id' => $eqLogic->getId(), 'eqlogic_name' => $eqLogic->getName(), 'cmd_name' => $this->getName(), 'reload' => $reload));
 	}
 
 	private function eventTopicMismatch() {
 		$eqLogic = $this->getEqLogic();
-		$eqLogic->log('warning', 'Le topic de la commande ' . $this->getLogName() .
-			" est incompatible du topic de l'équipement associé");
-
+		$eqLogic->log('warning', sprintf(__("Le topic de la commande #%s# est incompatible du topic de l'équipement associé", __FILE__), $this->getHumanName()));
 		// Advise the desktop page (jMQTT.js) of the topic mismatch
-		event::add('jMQTT::cmdTopicMismatch',
-			array('eqlogic_name' => $eqLogic->getName(), 'cmd_name' => $this->getName()));
+		event::add('jMQTT::cmdTopicMismatch', array('eqlogic_name' => $eqLogic->getName(), 'cmd_name' => $this->getName()));
 	}
 
 	/**
@@ -118,15 +112,11 @@ class jMQTTCmd extends cmd {
 		}
 		$eqLogic = $this->getEqLogic();
 		$eqLogic->checkAndUpdateCmd($this, $value);
-		$eqLogic->log('info', '-> ' . $this->getLogName() . ' ' . $value);
-
+		$eqLogic->log('info', sprintf(__("Cmd #%s# <- %s", __FILE__), $this->getHumanName(), $value));
 		if ($this->isBattery() && !in_array($value[0], ['{','[',''])) {
-			if ($this->getSubType() == 'binary') {
-				$eqLogic->batteryStatus($value ? 100 : 10);
-			} else {
-				$eqLogic->batteryStatus($this->getCache('value'));
-			}
-			$eqLogic->log('info', '-> Update battery status');
+			$value = ($this->getSubType() == 'binary') ? ($value ? 100 : 10) : $this->getCache('value');
+			$eqLogic->batteryStatus($value);
+			$eqLogic->log('info', sprintf(__("Eq  #%s# <- Batterie à %s%%", __FILE__), $eqLogic->getHumanName(), $value));
 		}
 	}
 
@@ -136,32 +126,27 @@ class jMQTTCmd extends cmd {
 	 * @param array $jsonArray associative array
 	 */
 	public function updateJsonCmdValue($jsonArray) {
-
 		// if dependancy is not yet installed, we avoid issue when someone create some command with JSONPath
 		if (!class_exists('JsonPath\JsonObject')) {
-			$this->getEqLogic()->log('error', 'JsonPath-PHP library is not available. Please reinstall jMQTT dependancy.');
+			$this->getEqLogic()->log('error', __("La bibliothèque JsonPath-PHP n'a pas été trouvée. Merci de réinstaller les dépendances de jMQTT.", __FILE__));
 			return;
 		}
-
 		// Create JsonObject for JsonPath
 		$jsonobject=new JsonPath\JsonObject($jsonArray);
-
 		// Get and prepare the jsonPath
 		$jsonPath = $this->getJsonPath();
 		if ($jsonPath == '') return;
 		if ($jsonPath[0] != '$') $jsonPath = '$' . $jsonPath;
-
 		try {
 			$value = $jsonobject->get($jsonPath);
 			if ($value !== false)
 				$this->updateCmdValue(json_encode((count($value) > 1) ? $value : $value[0], JSON_UNESCAPED_SLASHES | JSON_UNESCAPED_UNICODE));
 			else
-				$this->getEqLogic()->log('info', 'valeur de la commande ' . $this->getLogName() . ' non trouvée');
+				$this->getEqLogic()->log('info', sprintf(__("Valeur de la commande #%s# non trouvée", __FILE__), $this->getHumanName()));
 		}
 		catch (Throwable $e) {
-			$this->getEqLogic()->log('error', 'Chemin JSON de la commande ' . $this->getLogName() . ' incorrect : "' . $this->getJsonPath() . '"');
+			$this->getEqLogic()->log('error', sprintf(__("Chemin JSON de la commande #%s# incorrect : '%s'", __FILE__), $this->getHumanName(), $this->getJsonPath()));
 		}
-
 	}
 
 	/**
@@ -174,8 +159,7 @@ class jMQTTCmd extends cmd {
 		if (is_array($jsonArray) && json_last_error() == JSON_ERROR_NONE)
 			return $jsonArray;
 		else {
-			$this->getEqLogic()->log('warning', 'Erreur de format JSON sur la commande ' . $this->getLogName() .
-				': ' . json_last_error_msg());
+			$this->getEqLogic()->log('warning', sprintf(__("Erreur de format JSON sur la commande #%s#: %s", __FILE__), $this->getHumanName(), json_last_error_msg()));
 			return null;
 		}
 	}
@@ -193,15 +177,12 @@ class jMQTTCmd extends cmd {
 	 * This method is called when a command is executed
 	 */
 	public function execute($_options = null) {
-
 		if ($this->getType() != 'action')
 			return;
-
 		$request = $this->getConfiguration('request', "");
 		$topic = $this->getTopic();
 		$qos = $this->getConfiguration('Qos', 1);
 		$retain = $this->getConfiguration('retain', 0);
-
 		switch ($this->getSubType()) {
 			case 'slider':
 				$request = str_replace('#slider#', $_options['slider'], $request);
@@ -220,10 +201,8 @@ class jMQTTCmd extends cmd {
 				$request = str_replace('#select#', $_options['select'], $request);
 				break;
 		}
-
 		$request = jeedom::evaluateExpression($request);
 		$this->getEqLogic()->publish($this->getEqLogic()->getName(), $topic, $request, $qos, $retain);
-
 		return $request;
 	}
 
@@ -234,19 +213,16 @@ class jMQTTCmd extends cmd {
 
 		foreach(array('request') as $key) {
 			$conf = $this->getConfiguration($key);
-
 // TODO: DELETEME: fix reached Jeedom Core stable since 4.2.7
 			// Add/remove special char before JSON starting by '{' because Jeedom Core breaks integer, boolean and null values
 			// https://github.com/jeedom/core/pull/1825
 			// https://github.com/jeedom/core/pull/1829
 			if (is_string($conf) && strlen($conf) >= 1 && $conf[0] == chr(6)) $this->setConfiguration($key, substr($conf, 1));
 // End of DELETEME
-
 			// If request is an array, it means a JSON (starting by '{') has been parsed in 'request' field (parsed by getValues in jquery.utils.js)
 			if (is_array($conf) && (($conf = json_encode($conf, JSON_UNESCAPED_UNICODE)) !== FALSE))
 				$this->setConfiguration($key, $conf);
 		}
-
 		// Specific command : status for Broker eqpt
 		if ($this->getLogicalId() == jMQTT::CLIENT_STATUS && $this->getEqLogic()->getType() == jMQTT::TYP_BRK) {
 			if (!isset($this->name)) $this->setName(jMQTT::CLIENT_STATUS);
@@ -348,18 +324,15 @@ class jMQTTCmd extends cmd {
 			// If retain mode changed
 			if ($this->_preSaveInformations['retain'] != $this->getConfiguration('retain', 0)) {
 
-				$cmdLogName = $this->getLogName();
-
 				// It's enabled now
 				if ($this->getConfiguration('retain', 0)) {
-					$eqLogic->log('info', $cmdLogName . ': mode retain activé');
-				}
-				else{
+					$eqLogic->log('info', sprintf(__("Mode retain activé sur la commande #%s#", __FILE__), $this->getHumanName()));
+				} else {
 					//If broker eqpt is enabled
 					if ($eqLogic->getBroker()->getIsEnable()) {
 						// A null payload should be sent to the broker to erase the last retained value
 						// Otherwise, this last value remains retained at broker level
-						$eqLogic->log('info', $cmdLogName . ': mode retain désactivé, efface la dernière valeur mémorisée sur le broker');
+						$eqLogic->log('info', sprintf(__("Mode retain désactivé sur la commande #%s#, effacement de la dernière dand le Broker", __FILE__), $this->getHumanName()));
 						$eqLogic->publish($eqLogic->getName(), $this->getTopic(), '', 1, 1);
 					}
 				}
@@ -454,14 +427,14 @@ class jMQTTCmd extends cmd {
 	public function preRemove() {
 		$eqLogic = $this->getEqLogic();
 		if ($eqLogic) {
-			$eqLogic->log('info', 'Removing command ' . $this->getLogName());
+			$eqLogic->log('info', sprintf(__("Suppression de la commande #%s#", __FILE__), $eqLogic->getHumanName()));
 			// Remove batteryStatus from eqLogic on delete
 			if ($this->isBattery()) {
 				$eqLogic->setStatus('battery', null);
 				$eqLogic->setStatus('batteryDatetime', null);
 			}
 		} else {
-			jMQTT::logger('info', 'Removing orphan command #'.$this->getId().'#');
+			jMQTT::logger('info', sprintf(__("Suppression de la commande orpheline #%s#", __FILE__), $this->getId()));
 		}
 		$listener = listener::searchClassFunctionOption(__CLASS__, 'listenerAction', '"cmd":"'.$this->getId().'"');
 		foreach ($listener as $l) {
@@ -586,15 +559,6 @@ class jMQTTCmd extends cmd {
 	 */
 	public function isJson() {
 		return $this->getJsonPath() != '';
-	}
-
-	/**
-	 * Return the name of this command for logging purpose
-	 * (basically, return concatenation of the eqLogic name and the cmd name)
-	 * @return string
-	 */
-	private function getLogName() {
-		return $this->getEqLogic()->getName()  . '|' . $this->getName();
 	}
 
 	/**